Move the code about API to the new apps
