+++ rurusetto/wiki/views.py
@@ -3,9 +3,7 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404, resolve_url
 from django.templatetags.static import static
 from django.utils import timezone
-from django.views.decorators.csrf import csrf_exempt
 from django.http import HttpResponse, JsonResponse
-from .serializers import RulesetSerializer
 from .models import Changelog, Ruleset, Subpage, RecommendBeatmap, Action, RulesetStatus
 from users.models import Profile
 from django.contrib.auth.models import User
@@ -830,39 +828,3 @@ def redirect_from_old_link(request, slug):
         return redirect('wiki', slug=slug)
     else:
         return HttpResponse(status=404)
-
-
-# Views for API
-
-
-@csrf_exempt
-def ruleset_list(request):
-    """
-    View for return the ruleset in JSON format.
-
-    :param request: WSGI request from user
-    :return: All ruleset in website with its metadata in JSON format.
-    """
-    if request.method == 'GET':
-        rulesets = Ruleset.objects.all()
-        serializer = RulesetSerializer(rulesets, many=True)
-        return JsonResponse(serializer.data, safe=False)
-
-
-@csrf_exempt
-def ruleset_detail(request, slug):
-    """
-    View for return the specific ruleset that user pass by using its slug in JSON format.
-
-    :param request: WSGI request from user
-    :return: Specific ruleset metadata in JSON format.
-    """
-    # try to fetch ruleset from database
-    try:
-        ruleset = Ruleset.objects.get(slug=slug)
-    except Ruleset.DoesNotExist:
-        return HttpResponse(status=404)
-
-    if request.method == 'GET':
-        serializer = RulesetSerializer(ruleset)
-        return JsonResponse(serializer.data)
